superagi/tools/FinnhubSuperAgi/requirements.txt
finnhub_python==2.4.18
pydantic==1.10.8
superagi_tools==1.0.6
